+++ premier-design/widgets/layout/seo/utils/generateStructuredData.ts
@@ -1,9 +1,11 @@
+import {getFullCanonicalUrl} from './getFullCanonicalUrl';
+
 export const generateStructuredData = () => ({
     "@context": "https://schema.org",
     "@type": "Organization",
     "name": "Premium Interior",
     "url": "https://premium-interior.by",
-    "logo": "/logo.png",
+    "logo": getFullCanonicalUrl('/logo.png'),
     "sameAs": ["https://t.me/PremiumInterior", "https://www.instagram.com/proremont_zhl", "https://vk.com/premium_interior_zhl" ],
     "address": {
         "@type": "PostalAddress",